#12 rename: store 폴더 경로 변경 반영
