+++ src/main/java/com/codepatissier/keki/post/service/PostService.java
@@ -4,8 +4,8 @@ import com.codepatissier.keki.common.BaseException;
 import com.codepatissier.keki.post.dto.GetStorePostsRes;
 import com.codepatissier.keki.post.entity.PostImg;
 import com.codepatissier.keki.post.repository.PostRepository;
-import com.codepatissier.keki.stores.entity.Store;
-import com.codepatissier.keki.stores.repository.StoreRepository;
+import com.codepatissier.keki.store.entity.Store;
+import com.codepatissier.keki.store.repository.StoreRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
